# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Basic Banking System - OpenAPI 3.0",
    "description": "This is a Basic Banking System API based on the OpenAPI 3.0 specification.",
    "contact": {
      "email": "habibalfrb@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1"
    }
  ],
...
  "paths": {
...
    "/accounts": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Retrieve a list of accounts",
        "description": "Retrieve all registered accounts",
        "operationId": "getAccounts",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "format": "int32"
                          },
                          "userId": {
                            "type": "integer",
                            "format": "int32"
                          },
                          "bankName": {
                            "type": "string"
                          },
                          "bankAccountNumber": {
                            "type": "string"
                          },
                          "balance": {
                            "type": "integer",
                            "format": "int32"
                          }
                        }
                      }
                    }
                  },
                  "example": [
                    {
                      "id": 1,
                      "userId": 1,
                      "bankName": "BCA",
                      "bankAccountNumber": 2318792138671238,
                      "balance": 1500000
                    },
                    {
                      "id": 2,
                      "userId": 2,
                      "bankName": "Mandiri",
                      "bankAccountNumber": 121763127781238,
                      "balance": 95000000
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Failed to get all account data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Failed to get all account data"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Create a new bank account",
        "description": "Register a new bank account",
        "operationId": "createAccount",
        "requestBody": {
          "description": "Create a new bank account in banking system",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "integer",
                    "format": "int32",
                    "example": 1
                  },
                  "bankName": {
                    "type": "string",
                    "example": "BCA"
                  },
                  "bankAccountNumber": {
                    "type": "string",
                    "example": 2318792138671238
                  },
                  "balance": {
                    "type": "integer",
                    "format": "int32",
                    "example": 1500000
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "New bank account with account number 2318792138671238 has successfully added to user with id 1"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "format": "int32",
                          "example": 1
                        },
                        "userId": {
                          "type": "integer",
                          "format": "int32",
                          "example": 1
                        },
                        "bankName": {
                          "type": "string",
                          "example": "BCA"
                        },
                        "bankAccountNumber": {
                          "type": "string",
                          "example": 2318792138671238
                        },
                        "balance": {
                          "type": "integer",
                          "format": "int32",
                          "example": 1500000
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bank account already exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Bank account already exist"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to create bank account",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Failed to create a new account"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{id}": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Retrieve account by id",
        "description": "Retrieve account information detail",
        "operationId": "getAccountById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of account to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "format": "int32",
                          "example": 1
                        },
                        "userId": {
                          "type": "integer",
                          "format": "int32",
                          "example": 1
                        },
                        "bankName": {
                          "type": "string",
                          "example": "BCA"
                        },
                        "bankAccountNumber": {
                          "type": "string",
                          "example": 2318792138671238
                        },
                        "balance": {
                          "type": "integer",
                          "format": "int32",
                          "example": 1500000
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid account id format",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Account ID must be a valid number"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Account does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Bank account does not exist"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed get account data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Failed to get account data"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/authenticate": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Authenticate",
        "description": "Authenticate access token",
        "operationId": "authUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully authenticated"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "format": "int32",
                          "example": 2
                        },
                        "name": {
                          "type": "string",
                          "example": "Kathy Joannet"
                        },
                        "email": {
                          "type": "string",
                          "example": "kjoannet0@godaddy.com"
                        },
                        "profile": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer",
                              "format": "int32",
                              "example": 2
                            },
                            "identityType": {
                              "type": "string",
                              "example": "SIM"
                            },
                            "identityNumber": {
                              "type": "string",
                              "example": 2517309437874406
                            },
                            "address": {
                              "type": "string",
                              "example": "325 Bartillon Way"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}